core: move build version reading to its own package
This fixes the build with Go 1.17, which does not have
BuildInfo.Settings yet.
